requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# MiniProj/portscan.py: 1
python_nmap == 0.6.1
